Rearrange project structure and sintax format
