+++ package.json
@@ -10,6 +10,8 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "babel-plugin-module-resolver": "^4.0.0",
+    "module-resolver": "^1.0.0",
     "react": "16.9.0",
     "react-native": "0.61.5",
     "react-native-elements": "^1.2.7",
