+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import AppContainer from './navigators';
+import AppContainer from 'navigators';
 
 const App = () => (
   <>
