--- src/components/common/Header.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import {Text, StyleSheet} from 'react-native';
-
-const Header = ({title}) => <Text style={styles.title}> {title} </Text>;
-
-const styles = StyleSheet.create({
-  title: {
-    fontWeight: 'bold',
-    fontSize: 20,
-    alignSelf: 'center',
-    marginTop: 100,
-  },
-});
-
-export {Header};
