+++ src/components/common/Header/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import { Text } from 'react-native';
+import styles from './styles';
+
+const Header = ({ title }) => <Text style={styles.title}> {title} </Text>;
+
+export default Header;
