+++ src/components/common/Header/styles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  title: {
+    alignSelf: 'center',
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginTop: 100,
+  },
+});
+
+export default styles;
