+++ src/components/common/InputForm/styles.js
@@ -1,0 +1,26 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    margin: 15,
+  },
+
+  input: {
+    borderColor: 'black',
+    borderWidth: 1,
+    fontSize: 16,
+    height: 40,
+    padding: 5,
+    textAlign: 'center',
+    width: 200,
+  },
+
+  title: {
+    margin: 5,
+    textTransform: 'uppercase',
+  },
+});
+
+export default styles;
