--- src/components/common/index.js
@@ -1,4 +1,0 @@
-export * from './Header';
-export * from './InputForm';
-export * from './ButtonForm';
-export * from './Spacer';
