+++ src/constants/colors.js
@@ -1,0 +1,2 @@
+export const BLACK = '#000000';
+export const WHITE = '#FFFFFF';
