+++ src/navigators/index.js
@@ -1,7 +1,7 @@
-import {createAppContainer} from 'react-navigation';
-import {createStackNavigator} from 'react-navigation-stack';
+import { createAppContainer } from 'react-navigation';
+import { createStackNavigator } from 'react-navigation-stack';
 
-import LoginScreen from '../screen/LoginScreen';
+import LoginScreen from 'screens/LoginScreen';
 
 const AppNavigator = createStackNavigator(
   {
